feat : 포인트 거래내역 남은 포인트 기능

+++ src/main/java/avengers/lion/item/service/ItemService.java
@@ -51,7 +51,7 @@ public class ItemService {
         if(item.getStockCount()<count){
             throw new BusinessException(ExceptionType.STOCK_NOT_AVAILABLE);
         }
-        member.buyItemByPoint(price);
+        member.buyItemByPoint((long) price);
         walletService.addPointTransaction(price, member);
         item.buyItem(count);
